Add a new trtllm_quant.py to manage TRT-LLM-backed quantized linear ops
